Disconnection detection fix.  Increase fread() size to 1 GiB.
* Due to the way empty() works, \Net\Stream\Client would incorrectly
detect that a socket was disconnected if a single ascii '0' was
received.  Changed to use strlen() instead.

* Increased fread() read size to a gigabyte in the case of extremely
large data chucks that could be realistically buffered, at least far
beyond the original 8KiB.  I see no reason why this value should not be
set to a unrealistically large number.  While it's unlikely the OS
would buffer a solid GiB, several megabytes could be realistic.

* Also, removed unused $rawSocket member.